pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "pixelart-map"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0",
    "httpx>=0.27",
    "tqdm>=4.66",
    "flask>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]
claude = [
    "anthropic>=0.40",
]

[project.scripts]
pixelart-analyze = "pixelart_map.analyzer:main"
pixelart-browse = "pixelart_map.browser:main"

[tool.setuptools.packages.find]
include = ["pixelart_map*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
